+++ packages/metascraper-author/index.js
@@ -1,12 +1,9 @@
 'use strict'
 
-const { isUrl, titleize } = require('@metascraper/helpers')
+const { getValue, isUrl, titleize } = require('@metascraper/helpers')
 const { isString } = require('lodash')
 
-const REGEX_BY = /^[\s\n]*by|@[\s\n]*/im
 const REGEX_STRICT = /^\S+\s+\S+/
-
-const removeBy = value => value.replace(REGEX_BY, '')
 
 /**
  * Wrap a rule with validation and formatting logic.
@@ -17,7 +14,10 @@ const removeBy = value => value.replace(REGEX_BY, '')
 
 const wrap = rule => ({ htmlDom }) => {
   const value = rule(htmlDom)
-  return isString(value) && !isUrl(value, {relative: false}) && titleize(removeBy(value))
+
+  return isString(value) &&
+  !isUrl(value, {relative: false}) &&
+  titleize(value, {removeBy: true})
 }
 
 /**
@@ -29,19 +29,8 @@ const wrap = rule => ({ htmlDom }) => {
 
 const strict = rule => $ => {
   const value = rule($)
-  if (!REGEX_STRICT.test(value)) return
-  return value
+  return REGEX_STRICT.test(value) && value
 }
-
-const getFirst = ($, collection) =>
-  collection
-    .filter((i, el) =>
-      $(el)
-        .text()
-        .trim()
-    )
-    .first()
-    .text()
 
 /**
  * Rules.
@@ -53,57 +42,15 @@ module.exports = () => ({
     wrap($ => $('meta[property="article:author"]').attr('content')),
     wrap($ => $('meta[name="author"]').attr('content')),
     wrap($ => $('meta[name="sailthru.author"]').attr('content')),
-    wrap($ =>
-      $('[rel="author"]')
-        .first()
-        .text()
-    ),
-    wrap($ =>
-      $('[itemprop*="author"] [itemprop="name"]')
-        .first()
-        .text()
-    ),
-    wrap($ =>
-      $('[itemprop*="author"]')
-        .first()
-        .text()
-    ),
+    wrap($ => getValue($, $('[rel="author"]'))),
+    wrap($ => getValue($, $('[itemprop*="author"] [itemprop="name"]'))),
+    wrap($ => getValue($, $('[itemprop*="author"]'))),
     wrap($ => $('meta[property="book:author"]').attr('content')),
-    strict(
-      wrap($ =>
-        $('a[class*="author"]')
-          .first()
-          .text()
-      )
-    ),
-    strict(
-      wrap($ =>
-        $('[class*="author"] a')
-          .first()
-          .text()
-      )
-    ),
-    strict(wrap($ => getFirst($, $('a[href*="/author/"]')))),
-    wrap($ =>
-      $('a[class*="screenname"]')
-        .first()
-        .text()
-    ),
-    strict(
-      wrap($ =>
-        $('[class*="author"]')
-          .first()
-          .text()
-      )
-    ),
-    strict(
-      wrap($ =>
-        $('[class*="byline"]')
-          .first()
-          .text()
-      )
-    ),
-    wrap($ => getFirst($, $('.fullname'))),
-    wrap($ => $('[class*="user-info"]').text())
+    strict(wrap($ => getValue($, $('a[class*="author"]')))),
+    strict(wrap($ => getValue($, $('[class*="author"] a')))),
+    strict(wrap($ => getValue($, $('a[href*="/author/"]')))),
+    wrap($ => getValue($, $('a[class*="screenname"]'))),
+    strict(wrap($ => getValue($, $('[class*="author"]')))),
+    strict(wrap($ => getValue($, $('[class*="byline"]'))))
   ]
 })