+++ packages/metascraper-helpers/index.js
@@ -1,19 +1,21 @@
 'use strict'
 
+const { trim, flow, isEmpty } = require('lodash')
 const condenseWhitespace = require('condense-whitespace')
 const isRelativeUrl = require('is-relative-url')
 const { resolve: resolveUrl } = require('url')
 const sanetizeUrl = require('normalize-url')
 const smartquotes = require('smartquotes')
-const { flow, isNil } = require('lodash')
 const toTitle = require('to-title-case')
 const urlRegex = require('url-regex')
 
-const isUrl = (url, {relative = true} = {}) => {
-  if (isNil(url)) return false
-  if (!relative) return urlRegex().test(url)
-  return isRelativeUrl(url) || urlRegex().test(url)
-}
+const REGEX_BY = /^[\s\n]*by|@[\s\n]*/i
+
+const urlTest = (url, {relative = true}) => relative
+  ? isRelativeUrl(url) || urlRegex().test(url)
+  : urlRegex().test(url)
+
+const isUrl = (url, opts = {}) => !isEmpty(url) && urlTest(url, opts)
 
 const normalizeUrl = url => sanetizeUrl(url, { stripWWW: false })
 
@@ -27,11 +29,25 @@ const getUrl = (baseUrl, relativePath) => (
   normalizeUrl(getAbsoluteUrl(baseUrl, relativePath))
 )
 
+const removeByPrefix = flow([
+  value => value.replace(REGEX_BY, ''),
+  trim
+])
+
 const createTitle = flow([condenseWhitespace, smartquotes])
 
-const titleize = (src, { capitalize = false } = {}) => {
-  const title = createTitle(src)
-  return capitalize ? toTitle(title) : title
+const titleize = (src, { capitalize = false, removeBy = false } = {}) => {
+  let title = createTitle(src)
+  if (removeBy) title = removeByPrefix(title).trim()
+  if (capitalize) title = toTitle(title)
+  return title
+}
+
+const defaultFn = el => el.text().trim()
+
+const getValue = ($, collection, fn = defaultFn) => {
+  const el = collection.filter((i, el) => fn($(el))).first()
+  return fn(el)
 }
 
 module.exports = {
@@ -39,5 +55,6 @@ module.exports = {
   getUrl,
   isUrl,
   normalizeUrl,
-  getAbsoluteUrl
+  getAbsoluteUrl,
+  getValue
 }