+++ packages/metascraper/src/get-data.js
@@ -1,24 +1,25 @@
 'use strict'
 
-const { reduce, isEmpty } = require('lodash')
+const { isEmpty } = require('lodash')
+const pReduce = require('p-reduce')
 
-const getValue = ({ htmlDom, url, conditions, meta }) => {
+const getValue = async ({ htmlDom, url, conditions, meta }) => {
   const size = conditions.length
   let index = -1
   let value
 
   while (isEmpty(value) && index++ < size - 1) {
-    value = conditions[index]({ htmlDom, url, meta })
+    value = await conditions[index]({ htmlDom, url, meta })
   }
 
   return value
 }
 
 const getData = ({ rules, htmlDom, url }) =>
-  reduce(
+  pReduce(
     rules,
-    (acc, [propName, conditions]) => {
-      const value = getValue({ htmlDom, url, conditions, meta: acc })
+    async (acc, [propName, conditions]) => {
+      const value = await getValue({ htmlDom, url, conditions, meta: acc })
       acc[propName] = !isEmpty(value) ? value : null
       return acc
     },