+++ packages/metascraper/src/index.js
@@ -1,20 +1,25 @@
 'use strict'
 
 const { isUrl } = require('@metascraper/helpers')
-const { isEmpty } = require('lodash')
+const { isEmpty, partial } = require('lodash')
 
-const loadRules = require('./load-rules')
+const {autoload: autoLoadRules, load: loadRules} = require('./load-rules')
 const loadHTML = require('./load-html')
 const getData = require('./get-data')
 
-module.exports = async ({ url, html } = {}) => {
-  const rules = await loadRules()
+const create = loader => {
+  return async ({url, html} = {}) => {
+    const rules = await loader()
 
-  if (!isUrl(url)) throw new TypeError('You need to provide a valid url.')
-  if (isEmpty(html)) {
-    throw new TypeError('You need to provide a valid HTML markup.')
+    if (!isUrl(url)) throw new TypeError('You need to provide a valid url.')
+    if (isEmpty(html)) {
+      throw new TypeError('You need to provide a valid HTML markup.')
+    }
+
+    const htmlDom = loadHTML(html)
+    return getData({ rules, htmlDom, url })
   }
-
-  const htmlDom = loadHTML(html)
-  return getData({ rules, htmlDom, url })
 }
+
+module.exports = create(autoLoadRules)
+module.exports.load = rules => create(partial(loadRules, rules))